add cmake tools & fix reorder warnings

+++ src/common/QueueFamily.cpp
@@ -17,7 +17,7 @@ QueueFamilyIndices QueueFamily::FindQueueFamilies(const VkPhysicalDevice& device
 
   // Iterate through families until one that supports requirements is found
   bool found = false;
-  for (int i = 0; !found && i < families.size(); ++i) {
+  for (size_t i = 0; !found && i < families.size(); ++i) {
     const auto& family = families.at(i);
 
     // Check for graphics support
